requirements.txt
click==6.6
requests==2.11.1
PyYAML==3.12

# Testing
pytest>=3.0.3,<=4.0.0
requests-mock>=1.1.0,<2.0.0
